// eMaestroD.NgUI/src/app/Manage/Models/prod-discount.ts
export interface ProdDiscount {
  prodDiscID?: number;
  prodID?: number;
  offerID?: number;
  comID?: any;
  descr?: any;
  startQty?: number;
  maxQty?: number;
  maxAmount?: number;
  discount?: number;
  dtStart?: any;
  dtEnd?: any;
  active?: boolean;
  crtBy?: string;
  crtDate?: Date;
  modBy?: string;
  modDate?: Date;
  cstID?: number;
}